* Fixed GWT code which was broken when document ID stopped being between -INT_MAX and INT_MAX * Made GWT client code get attached to XWikiDocument in .xar file so it need not be copied in manually.

+++ xwiki-platform-realtime-editor/xwiki-platform-realtime-editor-client/src/main/java/fr/loria/score/client/ClientDTO.java
@@ -30,19 +30,19 @@ import java.io.Serializable;
  */
 public class ClientDTO implements Serializable {
     private int siteId;
-    private int esid;
+    private long esid;
     private Document document;
 
     public ClientDTO() {
     }
 
-    public ClientDTO(Document document, int siteId, int esid) {
+    public ClientDTO(Document document, int siteId, long esid) {
         this.siteId = siteId;
         this.esid = esid;
         this.document = document;
     }
 
-    public ClientDTO(String content, int siteId, int esid) {
+    public ClientDTO(String content, int siteId, long esid) {
         this(new PlainDocument(content), siteId, esid);
     }
 
@@ -61,11 +61,11 @@ public class ClientDTO implements Serializable {
         return this;
     }
 
-    public int getEditingSessionId() {
+    public long getEditingSessionId() {
         return esid;
     }
 
-    public ClientDTO setEditingSessionId(int esid) {
+    public ClientDTO setEditingSessionId(long esid) {
         this.esid = esid;
         return this;
     }
